glyph-extraction/pyproject.toml
[project]
name = "codex-glyph-extraction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cairosvg>=2.8.2",
    "pillow>=12.0.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "svgpathtools>=1.7.1",
]
